refactor: 카트 및 관리자 페이지 상태 좁히기
- App에서 사용되는 상태 및 함수를 각 페이지별로 이동
